+++ app/models/product.py
@@ -6,7 +6,7 @@ class Product(db.Model):
     name = db.Column(db.String(64), nullable=False)
     description = db.Column(db.Text)
     ptype = db.Column(db.Integer, default=0)
-    price = db.Column(db.Numeric, default=0)
+    price = db.Column(db.Float, default=0)
     pieces = db.Column(db.Integer, default=0)
     duration = db.Column(db.Integer, default=0)
     deleted = db.Column(db.Boolean, default=False)
